offers: add encoding and decoding for offer TLV
This commit begins with a minimal subset of TLVs required to
implement offers. Note that fields such as currency and blinded
paths are excluded at present.

Note: pubkeys are currently expressed as x-only, this is going to
change in the specification (likely), so we use our btcec.Pubkey
type over [32] so that it's easy to make the switch when required.