offers: add records from parsed types for tree calculation
When we decode an offer from another party, it's possible that it
contains odd TLVs that our code does not recognize. We need to include
these fields in our merkle root calculation, so this commit adds a
helper will will create tlv records that simply writes the bytes as
we originally read them out (in a tlv stream) for inclusion in root
calculation.