lnwire: add invoice type sub-namespace tlv type to decode payload
Bolt 12 invoices are transmitted in TLV type = 64 in onion message
payloads. So far, this code has been structured to place all final
hop payloads (>= 64) in a generic FinalPayloads slice for access
by higher layers. This assumes that these payloads will be _odd_
(since our decoding will fail for unknown even tlvs). This change
updates our decoding to simply read the invoice sub-namespace out
as a un-parsed set of bytes (so that we don't fail on the even tlv),
and leave higher-level decoding to other layers.