multi: move onion message path creation into routes package
Pull out all of this logic so that it can be re-used when it comes
time to create payments to blinded paths. So far as possible, this
is just a move commit, although we do introduce the concept of a
path request. Subsequent commits will update the structure to allow
for easier mocking.